+++ packages/blob/src/client.ts
@@ -5,7 +5,7 @@ import type { IncomingMessage } from 'node:http';
 // the `undici` module will be replaced with https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API
 // for browser contexts. See ./undici-browser.js and ./package.json
 import { fetch } from 'undici';
-import type { BlobCommandOptions } from './helpers';
+import type { BlobCommandOptions, WithUploadProgress } from './helpers';
 import { BlobError, getTokenFromOptionsOrEnv } from './helpers';
 import { createPutMethod } from './put';
 import type { PutBlobResult } from './put-helpers';
@@ -20,7 +20,7 @@ import { createCreateMultipartUploaderMethod } from './multipart/create-uploader
 // This types omits all options that are encoded in the client token.
 export interface ClientCommonCreateBlobOptions {
   /**
-   * Whether the blob should be publicly accessible. Support for private blobs is planned.
+   * Whether the blob should be publicly accessible.
    */
   access: 'public';
   /**
@@ -42,7 +42,9 @@ export interface ClientTokenOptions {
 }
 
 // shared interface for put and upload
-interface ClientCommonPutOptions extends ClientCommonCreateBlobOptions {
+interface ClientCommonPutOptions
+  extends ClientCommonCreateBlobOptions,
+    WithUploadProgress {
   /**
    * Whether to use multipart upload. Use this when uploading large files. It will split the file into multiple parts, upload them in parallel and retry failed parts.
    */
@@ -53,12 +55,6 @@ function createPutExtraChecks<
   TOptions extends ClientTokenOptions & ClientCommonCreateBlobOptions,
 >(methodName: string) {
   return function extraChecks(options: TOptions) {
-    if (typeof window === 'undefined') {
-      throw new BlobError(
-        `${methodName} must be called from a client environment`,
-      );
-    }
-
     if (!options.token.startsWith('vercel_blob_client_')) {
       throw new BlobError(`${methodName} must be called with a client token`);
     }
@@ -89,7 +85,7 @@ export const put = createPutMethod<ClientPutCommandOptions>({
 // vercelBlob. createMultipartUpload()
 // vercelBlob. uploadPart()
 // vercelBlob. completeMultipartUpload()
-// vercelBlob. createMultipartUploaded()
+// vercelBlob. createMultipartUploader()
 
 export type ClientCreateMultipartUploadCommandOptions =
   ClientCommonCreateBlobOptions & ClientTokenOptions;
@@ -110,7 +106,8 @@ export const createMultipartUploader =
 
 type ClientMultipartUploadCommandOptions = ClientCommonCreateBlobOptions &
   ClientTokenOptions &
-  CommonMultipartUploadOptions;
+  CommonMultipartUploadOptions &
+  WithUploadProgress;
 
 export const uploadPart =
   createUploadPartMethod<ClientMultipartUploadCommandOptions>({
@@ -160,12 +157,6 @@ export type UploadOptions = ClientCommonPutOptions & CommonUploadOptions;
 export const upload = createPutMethod<UploadOptions>({
   allowedOptions: ['contentType'],
   extraChecks(options) {
-    if (typeof window === 'undefined') {
-      throw new BlobError(
-        'client/`upload` must be called from a client environment',
-      );
-    }
-
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition -- Runtime check for DX.
     if (options.handleUploadUrl === undefined) {
       throw new BlobError(
@@ -258,7 +249,7 @@ async function verifyCallbackSignature({
   return verified;
 }
 
-function hexToArrayByte(input: string): ArrayBuffer {
+function hexToArrayByte(input: string): Buffer {
   if (input.length % 2 !== 0) {
     throw new RangeError('Expected string to be an even number of characters');
   }
@@ -384,8 +375,8 @@ export async function handleUpload({
       const signatureHeader = 'x-vercel-signature';
       const signature = (
         'credentials' in request
-          ? request.headers.get(signatureHeader) ?? ''
-          : request.headers[signatureHeader] ?? ''
+          ? (request.headers.get(signatureHeader) ?? '')
+          : (request.headers[signatureHeader] ?? '')
       ) as string;
 
       if (!signature) {
@@ -453,7 +444,8 @@ async function retrieveClientToken(options: {
 }
 
 function toAbsoluteUrl(url: string): string {
-  return new URL(url, window.location.href).href;
+  // location is available in web workers too: https://developer.mozilla.org/en-US/docs/Web/API/Window/location
+  return new URL(url, location.href).href;
 }
 
 function isAbsoluteUrl(url: string): boolean {
@@ -515,3 +507,5 @@ export interface GenerateClientTokenOptions extends BlobCommandOptions {
   addRandomSuffix?: boolean;
   cacheControlMaxAge?: number;
 }
+
+export { createFolder } from './create-folder';