+++ packages/blob/src/multipart/helpers.ts
@@ -15,7 +15,27 @@ export interface Part {
   etag: string;
 }
 
-export function toReadableStream(value: PutBody): ReadableStream<ArrayBuffer> {
+const supportsNewBlobFromArrayBuffer = new Promise<boolean>((resolve) => {
+  // React Native doesn't support creating a Blob from an ArrayBuffer, so we feature detect it
+  try {
+    const helloAsArrayBuffer = new Uint8Array([104, 101, 108, 108, 111]);
+    const blob = new Blob([helloAsArrayBuffer]);
+    blob
+      .text()
+      .then((text) => {
+        resolve(text === 'hello');
+      })
+      .catch(() => {
+        resolve(false);
+      });
+  } catch {
+    resolve(false);
+  }
+});
+
+export async function toReadableStream(
+  value: PutBody,
+): Promise<ReadableStream<ArrayBuffer | Uint8Array>> {
   // Already a ReadableStream, nothing to do
   if (value instanceof ReadableStream) {
     return value as ReadableStream<ArrayBuffer>;
@@ -32,18 +52,27 @@ export function toReadableStream(value: PutBody): ReadableStream<ArrayBuffer> {
     return Readable.toWeb(value) as ReadableStream<ArrayBuffer>;
   }
 
-  let streamValue: Uint8Array | ArrayBuffer;
+  let streamValue: Uint8Array;
 
+  // While ArrayBuffer is valid as a fetch body, when used in a ReadableStream it will fail in Node.js with
+  // The "chunk" argument must be of type string or an instance of Buffer or Uint8Array. Received an instance of ArrayBuffer
   if (value instanceof ArrayBuffer) {
+    streamValue = new Uint8Array(value);
+  } else if (isNodeJsBuffer(value)) {
     streamValue = value;
-  } else if (isNodeJsBufferOrString(value)) {
-    streamValue = value.buffer;
   } else {
-    streamValue = stringToUint8Array(value);
+    // value is a string, we need to convert it to a Uint8Array to get create a stream from it
+    streamValue = stringToUint8Array(value as string);
+  }
+
+  // This line ensures that even when we get a buffer of 70MB, we'll create a stream out of it so we can have
+  // better progress indication during uploads
+  if (await supportsNewBlobFromArrayBuffer) {
+    return new Blob([streamValue]).stream();
   }
 
   // from https://github.com/sindresorhus/to-readable-stream/blob/main/index.js
-  return new ReadableStream<ArrayBuffer>({
+  return new ReadableStream<ArrayBuffer | Uint8Array>({
     start(controller) {
       controller.enqueue(streamValue);
       controller.close();
@@ -52,7 +81,7 @@ export function toReadableStream(value: PutBody): ReadableStream<ArrayBuffer> {
 }
 
 // From https://github.com/sindresorhus/is-stream/
-function isNodeJsReadableStream(value: PutBody): value is Readable {
+export function isNodeJsReadableStream(value: PutBody): value is Readable {
   return (
     typeof value === 'object' &&
     typeof (value as Readable).pipe === 'function' &&
@@ -68,6 +97,6 @@ function stringToUint8Array(s: string): Uint8Array {
   return enc.encode(s);
 }
 
-function isNodeJsBufferOrString(input: Buffer | string): input is Buffer {
-  return isBuffer(input);
+function isNodeJsBuffer(value: PutBody): value is Buffer {
+  return isBuffer(value);
 }