+++ test/next/src/app/layout.tsx
@@ -9,7 +9,7 @@ export default function RootLayout({
   children,
 }: {
   children: React.ReactNode;
-}): JSX.Element {
+}): React.JSX.Element {
   return (
     <html lang="en">
       <body>{children}</body>