+++ test/next/test/@vercel/blob/blob.visual.test.ts
@@ -7,12 +7,13 @@ const prefix =
 test.describe('@vercel/blob', () => {
   test.describe('page', () => {
     test('serverless', async ({ page }) => {
-      // eslint-disable-next-line no-console -- [@vercel/style-guide@5 migration]
       console.log(`vercel/pages/blob/image?prefix=${prefix}`);
       const response = await page.goto(
         `vercel/pages/blob/image?prefix=${prefix}`,
       );
-      await expect(page).toHaveScreenshot('blob-image.png');
+      await expect(page.locator('#test-screenshot')).toHaveScreenshot(
+        'blob-image.png',
+      );
       expect(response?.status()).toBe(200);
     });
   });